feat: auth api
