+++ package.json
@@ -5,7 +5,7 @@
   "main": "src/index.ts",
   "scripts": {
     "build": "rm -rf dist && tsc -p tsconfig.json && tsc-alias -p tsconfig.json",
-    "start": "nodemon",
+    "start": "node dist/index.js",
     "dev": "nodemon",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
