+++ src/constants/product.ts
@@ -1,0 +1,4 @@
+export enum STATUS_PRODUCT {
+  active = "active",
+  inactive = "inactive",
+}
