+++ src/constants/user.ts
@@ -1,0 +1,19 @@
+export enum GENDER {
+  male = "male",
+  female = "female",
+  other = "other",
+}
+
+export enum STATUS_USER {
+  active = "active",
+  inactive = "inactive",
+  pending = "pending",
+}
+
+export enum ROLE {
+  user = "user",
+  admin = "admin",
+  staff = "staff",
+}
+
+
