+++ src/index.ts
@@ -3,8 +3,8 @@ import http from "http";
 import mongoose from "mongoose";
 import userRouter from "routes/user";
 import productRouter from "routes/product";
-import metadataRouter from "routes/meta_data";
-import orderRouter from "routes/order";
+import authRouter from "routes/auth";
+// import orderRouter from "routes/order";
 import { config } from "dotenv";
 import Log from "libraries/log";
 import cors from "cors";
@@ -58,10 +58,10 @@ app.use((req, res, next) => {
 });
 
 /* Routes */
-app.use("/user", userRouter);
-app.use("/product", productRouter);
-app.use("/meta_data", metadataRouter);
-app.use("/order", orderRouter);
+app.use("/users", userRouter);
+app.use("/products", productRouter);
+// app.use("/order", orderRouter);
+app.use("/auth", authRouter);
 
 app.get("/ping", (req, res: Response) => {
   res.status(200).json({
