+++ src/interfaces/basic.ts
@@ -1,0 +1,19 @@
+export type IModify = {
+  action: string;
+  date: string;
+};
+
+export type IOTP = {
+  code: string;
+  expired: Date;
+};
+
+export const Modify = {
+  action: String,
+  date: String,
+};
+
+export const OTP = {
+  code: String,
+  expired: Date,
+};
