+++ src/utils/token.ts
@@ -1,4 +1,3 @@
-import Log from "libraries/log";
 import jwt from "jsonwebtoken";
 import { Request } from "express";
